refactor(react_refresh): split scanner.rs into exports and components modules
