+++ crates/react_refresh/src/lib.rs
@@ -1,5 +1,7 @@
 #![doc = "Rust implementation of eslint-plugin-react-refresh rule logic."]
 
+mod components;
+mod exports;
 mod helpers;
 mod scanner;
